Created Prototype 3
Adding plane, and sphere, setup physics, and colliders, and made the sphere a Prefab.

My vision is to recreate the room at Summit One Vanderbilt in NYC which has a room full of floating metallic balls. People interacted with the balls, sending them flying around the mirrored room.

I want to recreate this, and have the balls spawn from the air, and when they reach the ground they dissapear.