Updated Spawner to rain spheres
So I had an issue where the spheres would dissapear as soon as they spawned basically. By removing the scale sphere foreach loop, the spheres would fall like rain.

Because the ground/plain was so small it seemed like the player was drowning in them. I thought about my original idea of them dissolving in the ground but that would make everything way less satisfying because ideally the player should be able to interact with them.

So, I made the ground/plane much larger and I like the result.

I would however, like to add more bounciness to the spheres but they are on max. I also would like to add more interactivity with them.