Explored the use of singletons
I explored singletons to manage spheres but ran into collider issues where spheres initially knocked the player over. After testing a sphere collider with Is Trigger enabled (which caused them to pass through the ground), I ultimately added a trigger collider to the player. This way, spheres still collided with the ground but didn’t physically push the player, and they could be removed on contact.