Added Score Variable in Singleton spawner and Score UI (TextMeshPro)
Implemented a score variable in the Spawner (singleton) so every time RemoveSphere() is called (when the player touches a sphere), the score increments.
Created a ScoreDisplay script using TextMeshPro UI that references Spawner.instance.score to show the updated score in the top-left corner.
Ran into the pink shader issue with TMP_Sub Mesh UI, but switched to LiberationSans SDF and and UnityAtlas material preset to fix it.
Now the score UI is properly displayed, and we can track how many spheres the player collects in real time.